+++ packages/spacecat-shared-data-access/src/models/index.d.ts
@@ -30,3 +30,4 @@ export type * from './site';
 export type * from './site-candidate';
 export type * from './site-top-page';
 export type * from './suggestion';
+export type * from './report';