+++ packages/spacecat-shared-data-access/src/models/index.js
@@ -31,3 +31,5 @@ export * from './site-candidate/index.js';
 export * from './site-top-page/index.js';
 export * from './site/index.js';
 export * from './suggestion/index.js';
+export * from './page-intent/index.js';
+export * from './report/index.js';