+++ packages/spacecat-shared-data-access/src/service/index.js
@@ -16,6 +16,7 @@ import { Service } from 'electrodb';
 
 import { instrumentAWSClient } from '@adobe/spacecat-shared-utils';
 import { EntityRegistry } from '../models/index.js';
+import { registerLogger } from '../util/logger-registry.js';
 
 export * from '../errors/index.js';
 export * from '../models/index.js';
@@ -58,6 +59,8 @@ const createElectroService = (client, config, log) => {
  * @returns {object} Data access collections for interacting with entities
  */
 export const createDataAccess = (config, log = console, client = undefined) => {
+  registerLogger(log);
+
   const rawClient = createRawClient(client);
   const electroService = createElectroService(rawClient, config, log);
   const entityRegistry = new EntityRegistry(electroService, config, log);