+++ packages/spacecat-shared-rum-api-client/src/functions/traffic-analysis.js
@@ -27,10 +27,12 @@ function getCWV(bundle, metric) {
   return measurements.length > 0 ? Math.max(...measurements) : null;
 }
 
-function containsEngagedScroll(bundle) {
-  return bundle.events
-    .some((e) => (e.checkpoint === 'viewmedia' || e.checkpoint === 'viewblock') && e.timeDelta >= 10000)
-    ? 1 : 0;
+function getLatestScroll(bundle) {
+  const scrolls = bundle.events
+    .filter((e) => e.checkpoint === 'viewmedia' || e.checkpoint === 'viewblock')
+    .map((e) => e.timeDelta);
+
+  return scrolls.length > 0 ? Math.max(...scrolls) : null;
 }
 
 function getNotFound(bundle) {
@@ -98,6 +100,7 @@ async function handler(bundles) {
     /* eslint-disable camelcase */
     const trafficData = trafficType(bundle, memo);
     const clicked = getClicked(bundle);
+    const latestScroll = getLatestScroll(bundle);
 
     return {
       path: new URL(bundle.url).pathname,
@@ -113,7 +116,8 @@ async function handler(bundles) {
       notfound: getNotFound(bundle),
       pageviews: bundle.weight,
       clicked,
-      engaged: containsEngagedScroll(bundle) || clicked,
+      engaged: (latestScroll >= 10000 || clicked) ? 1 : 0,
+      latest_scroll: latestScroll,
       lcp: getCWV(bundle, 'lcp'),
       inp: getCWV(bundle, 'inp'),
       cls: getCWV(bundle, 'cls'),