+++ packages/spacecat-shared-google-client/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [@adobe/spacecat-shared-google-client-v1.4.38](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.4.37...@adobe/spacecat-shared-google-client-v1.4.38) (2025-07-19)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#859](https://github.com/adobe/spacecat-shared/issues/859)) ([7ca9099](https://github.com/adobe/spacecat-shared/commit/7ca90994d61d07f71e580301365447b94ad07a52))
+
 # [@adobe/spacecat-shared-google-client-v1.4.37](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.4.36...@adobe/spacecat-shared-google-client-v1.4.37) (2025-07-12)